refactor: share context utilities and keywords
