+++ src/ai/databaseContextManager.ts
@@ -8,6 +8,11 @@ import type {
   MessageContext,
   ServerContext,
 } from "./contextManager.js";
+import {
+  extractTopicsFromMessage,
+  calculateRelevanceScore,
+  identifyContextualCues,
+} from "./contextUtils.js";
 
 /**
  * Database-backed Context Manager
@@ -58,14 +63,14 @@ export class DatabaseContextManager {
     const recentTopics = this.extractTopics(conversationHistory);
 
     // Calculate relevance score
-    const relevanceScore = this.calculateRelevanceScore(
+    const relevanceScore = calculateRelevanceScore(
       message,
       userProfile,
       channelContext
     );
 
     // Identify contextual cues
-    const contextualCues = this.identifyContextualCues(
+    const contextualCues = identifyContextualCues(
       message,
       conversationHistory
     );
@@ -95,7 +100,7 @@ export class DatabaseContextManager {
       serverContext,
       conversationHistory: conversationHistory.map((msg) => ({
         ...msg,
-        topics: this.extractTopicsFromMessage(msg.content),
+        topics: extractTopicsFromMessage(msg.content),
       })),
       relevanceScore,
       contextualCues,
@@ -239,370 +244,14 @@ export class DatabaseContextManager {
 
     for (const message of conversationHistory.slice(-10)) {
       // Last 10 messages
-      const messageTopics = this.extractTopicsFromMessage(message.content);
+      const messageTopics = extractTopicsFromMessage(message.content);
       messageTopics.forEach((topic) => topics.add(topic));
     }
 
     return Array.from(topics);
   }
 
-  /**
-   * Extract topics from a single message using keyword analysis
-   */
-  private extractTopicsFromMessage(content: string): string[] {
-    const topics: string[] = [];
-    const cleanContent = content.toLowerCase();
 
-    // Technology topics (Vietnamese + English)
-    const techKeywords = [
-      // Vietnamese
-      "công nghệ",
-      "lập trình",
-      "phần mềm",
-      "ứng dụng",
-      "website",
-      "web",
-      "app",
-      "máy tính",
-      "điện thoại",
-      "smartphone",
-      "laptop",
-      "pc",
-      "android",
-      "ios",
-      "trí tuệ nhân tạo",
-      "ai",
-      "machine learning",
-      "deep learning",
-      // English (still common in Vietnamese tech discussions)
-      "javascript",
-      "typescript",
-      "python",
-      "react",
-      "discord",
-      "bot",
-      "programming",
-      "code",
-      "coding",
-      "github",
-      "api",
-      "html",
-      "css",
-      "nodejs",
-    ];
-    techKeywords.forEach((keyword) => {
-      if (cleanContent.includes(keyword)) {
-        topics.push(`tech:${keyword}`);
-      }
-    });
-
-    // Gaming topics (Vietnamese + English)
-    const gameKeywords = [
-      // Vietnamese
-      "game",
-      "chơi game",
-      "gaming",
-      "trò chơi",
-      "game mobile",
-      "game online",
-      "liên minh huyền thoại",
-      "lol",
-      "pubg",
-      "free fire",
-      "valorant",
-      "fifa",
-      "pes",
-      "minecraft",
-      "roblox",
-      "among us",
-      // English gaming terms
-      "steam",
-      "xbox",
-      "playstation",
-      "nintendo",
-      "switch",
-      "ps5",
-      "ps4",
-    ];
-    gameKeywords.forEach((keyword) => {
-      if (cleanContent.includes(keyword)) {
-        topics.push(`gaming:${keyword}`);
-      }
-    });
-
-    // General interests (Vietnamese)
-    const interestKeywords = [
-      // Entertainment
-      "nhạc",
-      "âm nhạc",
-      "bài hát",
-      "ca sĩ",
-      "nhạc sĩ",
-      "kpop",
-      "vpop",
-      "phim",
-      "phim ảnh",
-      "điện ảnh",
-      "netflix",
-      "youtube",
-      "tiktok",
-      "sách",
-      "đọc sách",
-      "tiểu thuyết",
-      "truyện",
-      "manga",
-      "anime",
-      // Food & Lifestyle
-      "ăn",
-      "thức ăn",
-      "món ăn",
-      "nấu ăn",
-      "đồ ăn",
-      "quán ăn",
-      "nhà hàng",
-      "cafe",
-      "cà phê",
-      "trà",
-      "bánh",
-      "phở",
-      "bún",
-      "cơm",
-      "du lịch",
-      "travel",
-      "đi chơi",
-      "nghỉ dưỡng",
-      "vacation",
-      "nghệ thuật",
-      "art",
-      "vẽ",
-      "painting",
-      "nhiếp ảnh",
-      "photography",
-      // Sports & Health
-      "thể thao",
-      "bóng đá",
-      "football",
-      "tennis",
-      "badminton",
-      "cầu lông",
-      "gym",
-      "tập gym",
-      "fitness",
-      "yoga",
-      "chạy bộ",
-      "running",
-      "sức khỏe",
-    ];
-    interestKeywords.forEach((keyword) => {
-      if (cleanContent.includes(keyword)) {
-        topics.push(`interest:${keyword}`);
-      }
-    });
-
-    // Emotions/tone (Vietnamese)
-    const emotionKeywords = [
-      // Positive emotions
-      "vui",
-      "vui vẻ",
-      "hạnh phúc",
-      "happy",
-      "vui mừng",
-      "phấn khích",
-      "excited",
-      "yêu",
-      "thích",
-      "love",
-      "like",
-      "tuyệt vời",
-      "amazing",
-      "tốt",
-      "good",
-      "hài lòng",
-      "satisfied",
-      "thoải mái",
-      "comfortable",
-      "chill",
-      "relax",
-      // Negative emotions
-      "buồn",
-      "sad",
-      "khóc",
-      "cry",
-      "thất vọng",
-      "disappointed",
-      "tức giận",
-      "angry",
-      "giận",
-      "mad",
-      "bực mình",
-      "annoyed",
-      "lo lắng",
-      "worried",
-      "anxiety",
-      "stress",
-      "căng thẳng",
-      "nervous",
-      "mệt",
-      "tired",
-      "mệt mỏi",
-      "exhausted",
-      "chán",
-      "bored",
-      "bối rối",
-      "confused",
-      "hoang mang",
-      "lost",
-      "không hiểu",
-    ];
-    emotionKeywords.forEach((keyword) => {
-      if (cleanContent.includes(keyword)) {
-        topics.push(`emotion:${keyword}`);
-      }
-    });
-
-    // Vietnamese-specific topics
-    const vietnameseTopics = [
-      // Education
-      "học",
-      "học tập",
-      "study",
-      "trường",
-      "school",
-      "university",
-      "đại học",
-      "thi",
-      "exam",
-      "kiểm tra",
-      "test",
-      "bài tập",
-      "homework",
-      // Work
-      "làm việc",
-      "work",
-      "job",
-      "công việc",
-      "career",
-      "nghề nghiệp",
-      "công ty",
-      "company",
-      "office",
-      "văn phòng",
-      "meeting",
-      "họp",
-      // Daily life
-      "gia đình",
-      "family",
-      "bạn bè",
-      "friends",
-      "người yêu",
-      "boyfriend",
-      "girlfriend",
-      "thời tiết",
-      "weather",
-      "nóng",
-      "hot",
-      "lạnh",
-      "cold",
-      "mưa",
-      "rain",
-      "shopping",
-      "mua sắm",
-      "tiền",
-      "money",
-      "giá",
-      "price",
-    ];
-    vietnameseTopics.forEach((keyword) => {
-      if (cleanContent.includes(keyword)) {
-        topics.push(`vietnamese:${keyword}`);
-      }
-    });
-
-    return topics;
-  }
-
-  /**
-   * Calculate how relevant the message is for AI response
-   */
-  private calculateRelevanceScore(
-    message: ProcessedMessage,
-    userProfile: UserProfile,
-    channelContext: ChannelContext
-  ): number {
-    let score = 0.5; // Base score
-
-    // Higher score for mentions and replies
-    if (message.messageType === "mention") score += 0.3;
-    if (message.isReply) score += 0.2;
-
-    // Higher score for questions
-    if (message.cleanContent.includes("?")) score += 0.2;
-
-    // Higher score for users we've interacted with before
-    if (userProfile.interactionCount > 0) score += 0.1;
-
-    // Higher score for recent activity in channel
-    const timeSinceLastActivity = Date.now() - channelContext.lastActivity;
-    if (timeSinceLastActivity < 300000) score += 0.1; // Within 5 minutes
-
-    // Lower score for very short messages
-    if (message.cleanContent.length < 10) score -= 0.2;
-
-    // Higher score for messages with common interests
-    const messageTopics = this.extractTopicsFromMessage(message.cleanContent);
-    const commonTopics = messageTopics.filter(
-      (topic) =>
-        userProfile.recentTopics.includes(topic) ||
-        channelContext.recentTopics.includes(topic)
-    );
-    score += commonTopics.length * 0.1;
-
-    return Math.max(0, Math.min(1, score)); // Clamp between 0 and 1
-  }
-
-  /**
-   * Identify contextual cues that should influence the response
-   */
-  private identifyContextualCues(
-    message: ProcessedMessage,
-    conversationHistory: Array<{ content: string; author?: string }>
-  ): string[] {
-    const cues: string[] = [];
-
-    // Time-based cues
-    const hour = new Date().getHours();
-    if (hour < 6 || hour > 22) cues.push("late-night");
-    if (hour >= 6 && hour < 12) cues.push("morning");
-    if (hour >= 12 && hour < 17) cues.push("afternoon");
-    if (hour >= 17 && hour < 22) cues.push("evening");
-
-    // Message type cues
-    if (message.messageType === "mention") cues.push("mentioned");
-    if (message.isReply) cues.push("reply");
-
-    // Content cues
-    const content = message.cleanContent.toLowerCase();
-    if (content.includes("help") || content.includes("?"))
-      cues.push("needs-help");
-    if (content.includes("thanks") || content.includes("thank"))
-      cues.push("grateful");
-    if (content.includes("hello") || content.includes("hi"))
-      cues.push("greeting");
-    if (content.includes("bye") || content.includes("goodbye"))
-      cues.push("farewell");
-
-    // Conversation flow cues
-    if (conversationHistory.length === 0) cues.push("first-interaction");
-    if (conversationHistory.length > 5) cues.push("ongoing-conversation");
-
-    // Recent context cues
-    const recentMessages = conversationHistory.slice(-3);
-    if (recentMessages.some((msg) => msg.author === message.author.username)) {
-      cues.push("continuing-user");
-    }
-
-    return cues;
-  }
 
   /**
    * Update user profile with new interaction data
@@ -619,7 +268,7 @@ export class DatabaseContextManager {
     };
 
     // Extract and add topics to recent topics
-    const messageTopics = this.extractTopicsFromMessage(message.cleanContent);
+    const messageTopics = extractTopicsFromMessage(message.cleanContent);
     const newTopics = [...profile.recentTopics];
 
     messageTopics.forEach((topic) => {
@@ -705,7 +354,7 @@ export class DatabaseContextManager {
     updates.activeUsers = newActiveUsers.slice(-20);
 
     // Update recent topics
-    const messageTopics = this.extractTopicsFromMessage(message.cleanContent);
+    const messageTopics = extractTopicsFromMessage(message.cleanContent);
     const newTopics = [...context.recentTopics];
 
     messageTopics.forEach((topic) => {
@@ -803,7 +452,7 @@ export class DatabaseContextManager {
       content: msg.content,
       timestamp: msg.timestamp,
       author: msg.author || undefined,
-      topics: this.extractTopicsFromMessage(msg.content),
+      topics: extractTopicsFromMessage(msg.content),
     }));
   }
