+++ src/ai/topicKeywords.ts
@@ -1,0 +1,220 @@
+export const topicKeywords: Record<string, string[]> = {
+  tech: [
+    "công nghệ",
+    "lập trình",
+    "phần mềm",
+    "ứng dụng",
+    "website",
+    "web",
+    "app",
+    "máy tính",
+    "điện thoại",
+    "smartphone",
+    "laptop",
+    "pc",
+    "android",
+    "ios",
+    "trí tuệ nhân tạo",
+    "ai",
+    "machine learning",
+    "deep learning",
+    "javascript",
+    "typescript",
+    "python",
+    "react",
+    "discord",
+    "bot",
+    "programming",
+    "code",
+    "coding",
+    "github",
+    "api",
+    "html",
+    "css",
+    "nodejs",
+  ],
+  gaming: [
+    "game",
+    "chơi game",
+    "gaming",
+    "trò chơi",
+    "game mobile",
+    "game online",
+    "liên minh huyền thoại",
+    "lol",
+    "pubg",
+    "free fire",
+    "valorant",
+    "fifa",
+    "pes",
+    "minecraft",
+    "roblox",
+    "among us",
+    "steam",
+    "xbox",
+    "playstation",
+    "nintendo",
+    "switch",
+    "ps5",
+    "ps4",
+  ],
+  interest: [
+    "nhạc",
+    "âm nhạc",
+    "bài hát",
+    "ca sĩ",
+    "nhạc sĩ",
+    "kpop",
+    "vpop",
+    "phim",
+    "phim ảnh",
+    "điện ảnh",
+    "netflix",
+    "youtube",
+    "tiktok",
+    "sách",
+    "đọc sách",
+    "tiểu thuyết",
+    "truyện",
+    "manga",
+    "anime",
+    "ăn",
+    "thức ăn",
+    "món ăn",
+    "nấu ăn",
+    "đồ ăn",
+    "quán ăn",
+    "nhà hàng",
+    "cafe",
+    "cà phê",
+    "trà",
+    "bánh",
+    "phở",
+    "bún",
+    "cơm",
+    "du lịch",
+    "travel",
+    "đi chơi",
+    "nghỉ dưỡng",
+    "vacation",
+    "nghệ thuật",
+    "art",
+    "vẽ",
+    "painting",
+    "nhiếp ảnh",
+    "photography",
+    "thể thao",
+    "bóng đá",
+    "football",
+    "tennis",
+    "badminton",
+    "cầu lông",
+    "gym",
+    "tập gym",
+    "fitness",
+    "yoga",
+    "chạy bộ",
+    "running",
+    "sức khỏe",
+  ],
+  emotion: [
+    "vui",
+    "vui vẻ",
+    "hạnh phúc",
+    "happy",
+    "vui mừng",
+    "phấn khích",
+    "excited",
+    "yêu",
+    "thích",
+    "love",
+    "like",
+    "tuyệt vời",
+    "amazing",
+    "tốt",
+    "good",
+    "hài lòng",
+    "satisfied",
+    "thoải mái",
+    "comfortable",
+    "chill",
+    "relax",
+    "buồn",
+    "sad",
+    "khóc",
+    "cry",
+    "thất vọng",
+    "disappointed",
+    "tức giận",
+    "angry",
+    "giận",
+    "mad",
+    "bực mình",
+    "annoyed",
+    "lo lắng",
+    "worried",
+    "anxiety",
+    "stress",
+    "căng thẳng",
+    "nervous",
+    "mệt",
+    "tired",
+    "mệt mỏi",
+    "exhausted",
+    "chán",
+    "bored",
+    "bối rối",
+    "confused",
+    "hoang mang",
+    "lost",
+    "không hiểu",
+  ],
+  vietnamese: [
+    "học",
+    "học tập",
+    "study",
+    "trường",
+    "school",
+    "university",
+    "đại học",
+    "thi",
+    "exam",
+    "kiểm tra",
+    "test",
+    "bài tập",
+    "homework",
+    "làm việc",
+    "work",
+    "job",
+    "công việc",
+    "career",
+    "nghề nghiệp",
+    "công ty",
+    "company",
+    "office",
+    "văn phòng",
+    "meeting",
+    "họp",
+    "gia đình",
+    "family",
+    "bạn bè",
+    "friends",
+    "người yêu",
+    "boyfriend",
+    "girlfriend",
+    "thời tiết",
+    "weather",
+    "nóng",
+    "hot",
+    "lạnh",
+    "cold",
+    "mưa",
+    "rain",
+    "shopping",
+    "mua sắm",
+    "tiền",
+    "money",
+    "giá",
+    "price",
+  ],
+};
